+++ src/test/java/ru/ibs/appline/base/BaseTests.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import ru.ibs.appline.pegaobject.HomePage;
 
 import java.time.Duration;
