+++ src/test/java/ru/ibs/appline/pageobject/FormPage.java
@@ -1,8 +1,7 @@
-package ru.ibs.appline.pegaobject;
+package ru.ibs.appline.pageobject;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
