+++ src/test/java/ru/ibs/appline/pageobject/HomePage.java
@@ -1,6 +1,5 @@
-package ru.ibs.appline.pegaobject;
+package ru.ibs.appline.pageobject;
 
-import dev.failsafe.internal.util.Assert;
 import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
